bugfix: do computations in int range to avoid LED flickering due to wrapping

+++ src/pwm/led_fade/pwm_led_fade.c
@@ -25,7 +25,7 @@
 
 void on_pwm_wrap() {
 // this is the interrupt handler, called each time the PWM counter wraps
-    static uint16_t fade = 0;
+    static int fade = 0;
     static bool going_up = true;
     // Clear the interrupt flag that brought us here
     pwm_clear_irq(pwm_gpio_to_slice_num(PICO_DEFAULT_LED_PIN));
